+++ src/app/insight360/app/create/i360-create-list.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { DesignService } from '../shared/i360-design.service';
-import { Design } from '../../../../app_server/models/design';
+import { Design } from '../../../../../app_server/models/design';
 
 @Component({
   selector: 'sg-i360-create-list',
