+++ src/app/insight360/app/dashboard/i360-dashboard.component.css
