+++ src/app/insight360/app/i360-app.component.ts
@@ -1,0 +1,10 @@
+import { Component } from '@angular/core';
+
+@Component({
+  templateUrl: './i360-app.component.html',
+  styleUrls: ['./i360-app.component.css'],
+})
+
+export class Insight360AppComponent {
+
+}
