+++ src/app/insight360/app/shared/i360-design.service.ts
@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import { Design } from '../../../../app_server/models/design';
+import { Design } from '../../../../../app_server/models/design';
 
 @Injectable()
