+++ src/app/insight360/i360.component.ts
@@ -1,10 +1,10 @@
 import { Component } from '@angular/core';
 
+
 @Component({
   templateUrl: './i360.component.html',
-  styleUrls: ['./i360.component.css'],
 })
 
-export class Insight360AppComponent {
-
+export class Insight360Component {
+  
 }
